+++ lib/nretnil/compendium-api/content.rb
@@ -16,20 +16,20 @@ module Nretnil
       end
 
       def list(options = {})
-        defaults = { :Page => '1', :Count => '20', :'~Status' => ["deleted"].to_json }
+        defaults = { :Page => '1', :Count => '20' }
         query = defaults.merge(options)
-        response = @session.get( '/app/posts', query )
+        response = @session.get( '/api/posts', query )
       end
 
       def list_all(options = {})
         @content = []
         @options = options
-        response = { "Statistics" => { "Total" => 1 }}
+        response = { "stats" => { "total" => 1 }}
         page = 1
-        while response["Statistics"]["Total"] > @content.length
+        while response["stats"]["total"] > @content.length
           list_options = { :Count => '100', :Page => page }.merge(@options)
           response = list(list_options)
-          @content += response["Success"]
+          @content += response["posts"]
           page += 1
         end
         { 'Success' => @content }
@@ -40,36 +40,35 @@ module Nretnil
         response = @session.get( '/api/posts/' + post_id, query )
       end
 
-      def add(title,body,slug,publish_date,draft,options = {})
-        manditory = { :Title => title, :Body => body, :Slug => slug, :PublishDate => publish_date, :Draft => draft }
+      def add(user_id,title,body,slug,publish_date,content_type_id,options = {})
+        manditory = { :author_id => user_id, :post_title => title, :post_text => body, :url_lookup_token => slug, :publish_date => publish_date, :content_type_id => content_type_id }
         query = manditory.merge(options)
-        response = @session.post( '/app/post', query )
+        response = @session.post( '/api/posts', query.to_json )
+      end
+
+      def idea(title,options = {})
+        manditory = { :post_title => title }
+        query = manditory.merge(options)
+        response = @session.post( '/api/posts', query.to_json )
       end
 
       def update(post_id,options = {})
-        manditory = { :PostId => post_id }
+        manditory = { :id => post_id }
         query = manditory.merge(options)
-        response = @session.post( '/app/post', query )
+        response = @session.post( '/api/posts', query )
       end
 
-      def approve(post_ids,force = false)
-        request = []
-        post_ids.each do |post_id|
-          request << { "PostId" => post_id, "Operation" => "approve", "Notify" => "false", "Ping" => "false" }
-        end
-        response = @session.post( '/app/posts/moderate', {:Posts => request.to_json } )
+      def approve(post_ids = [],force = false)
+        response = @session.info('This Endpoint has been depreciated as of the New Workflow System')
       end
 
-      def decline(post_ids)
+      def take_down(post_id)
         request = []
-        post_ids.each do |post_id|
-          request << { "PostId" => post_id, "Operation" => "decline", "Notify" => "false", "Ping" => "false" }
-        end
-        response = @session.post( '/app/posts/moderate', {:Posts => request.to_json } )
+        response = @session.post( '/api/posts/' + post_id + '/takedown', {:Posts => request.to_json } )
       end
 
       def delete(post_id)
-        response = @session.delete( '/app/post/' + post_id )
+        response = @session.delete( '/api/posts/' + post_id )
       end
 
       def required_params