+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+<h4>0.10.1</h4>
+
+* Move post endpoint to use /api instead of /app to go along with new Compendium Workflow Release
+* Approve has been removed from the content class
+* Added idea creation for content class
+
 <h4>0.10.0</h4>
 
 * Added Callback Support